dom/base/test/test_parseHTMLUnsafe_from_expanded_principal.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1937488
-->
<head>
  <title>Test for Bug 1937488</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1937488">Mozilla Bug 1937488</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script>
const Cu = SpecialPowers.Cu;

let sandbox = Cu.Sandbox([window], {wantGlobalProperties: ["Document"]});
ok(
  Cu.evalInSandbox(`(() => {
    try {
      let doc = Document.parseHTMLUnsafe("");
    } catch (e) {}
    return true;
  })()`, sandbox),
  "parseHTMLUnsafe should not crash when used from an expanded principal",
);
</script>
</pre>
</body>
</html>
